- Frame extractor refactored

+++ Heedbook/HBLib/Utils/SftpClient.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
@@ -38,7 +38,7 @@ namespace HBLib.Utils
         public async Task UploadAsync(String localPath, String remotePath, String fileName)
         {
             await ConnectToSftpAsync();
-            using (var fs = new FileStream(localPath, FileMode.Open))
+            using (var fs = new FileStream(localPath, FileMode.Open, FileAccess.Read))
             {
                 _client.BufferSize = 4 * 1024;
                 _client.ChangeDirectory(_sftpSettings.DestinationPath + remotePath);
@@ -152,6 +152,28 @@ namespace HBLib.Utils
         }
 
         /// <summary>
+        ///     Download files to local disk.
+        /// </summary>
+        /// <param name="remotePath"></param>
+        /// <returns></returns>
+        public async Task<String> DownloadFromFtpToLocalDiskAsync(String remotePath, string pattern, String localPath = null)
+        {
+            await ConnectToSftpAsync();
+            Console.WriteLine("Successfully connected");
+            var filename = remotePath.Split('/').Last();
+
+            Console.WriteLine(localPath == null);
+            localPath = localPath == null ? localPath = _sftpSettings.DownloadPath + filename : localPath + filename;
+            Console.WriteLine($"{localPath}, {remotePath}");
+            using (var fs = File.OpenWrite(localPath))
+            {
+                await Task.Run(() => _client.DownloadFile(remotePath, fs));
+            }
+
+            return localPath;
+        }
+        
+        /// <summary>
         ///     Check file exists on server
         /// </summary>
         /// <param name="path">Specified directory + filename</param>
@@ -170,6 +192,31 @@ namespace HBLib.Utils
         public async Task DeleteFileIfExistsAsync(String path)
         {
             await ConnectToSftpAsync();
+            if (_client.Exists(path)) 
+                await Task.Run(() => _client.DeleteFile(path));conflict?name=Heedbook%252FHBOperations%252FHeedbook.sln&ancestor_oid=71a486bf139b886d6ad0fd7a39b03b29fc09da9f&base_oid=d6a55cb7d861f54f5e249d439ac87fdb245fb382&head_oid=29cccf1b9acab127d6b5a35d2533d65fca438522
+        }
+
+        /// <summary>
+        /// Lists all files in a directory using a pattern
+        /// </summary>
+        /// <param name="path">dir path in FTP</param>
+        /// <param name="patternToFind">pattern for filename</param>
+        /// <returns></returns>
+        public async Task<ICollection<string>> ListDirectoryFiles(string path, string patternToFind = null)
+        {
+            var result = new List<string>();
+            await ConnectToSftpAsync();
+
+            path = _sftpSettings.DestinationPath + path;
+
+            if (!_client.Exists(path))
+                return result;
+
+            if (patternToFind != null)
+                return _client.ListDirectory(path).Where(f => !f.IsDirectory && f.Name.Contains(patternToFind)).Select(f => f.Name).ToList();
+            else
+                return _client.ListDirectory(path).Where(f => !f.IsDirectory).Select(f => f.Name).ToList();          
+
             if (_client.Exists(_sftpSettings.DestinationPath + "/" + path))
             {
                 await Task.Run(() => _client.DeleteFile(_sftpSettings.DestinationPath + "/" + path));
